backend/requirements.txt
blinker==1.6.2
boto3==1.28.2
botocore==1.31.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.4
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
fonttools==4.41.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
progressbar2==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-utils==3.7.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sympy==1.12
threadpoolctl==3.1.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
